[CLOUDBEES-2296] Fixed the bogus "URI must start with a slash" problem
As you can see in the parseRequestParamters() method, the parsedParameters flag is getting set to true even without parsing the payload. This happens when the application calls methods like getParameterMap(), yet the payload is not of "application/x-www-form-urlencoded".

When such a request fails, Winstone fails to consume the rest of the payload, and if the connection is HTTP/1.1, next time when we read a request from this connection, we end up trying to read the request body as the HTTP request line, which fails.

I'm simplifying the code by using the inputData directly. Calling getInputStream() is harmful because it performs some checks (like whether the getReader() was already called.)