[JENKINS-66379] use default keystore format
The JKS keystore is insecure and is some environments not even allowed.

We do not care what format the keystore is in, so just use the default.

If someone has a non default keystore yet is using a keystore here then
this will break on upgrade, however it is unlikely that they would be
having a non default keystore as the normal reason for doing that is due
to a restriction like FIPS-140 and given the keystore we require has a
key using JDK here would be non compliant.